setup a basic hello world type app
I added some node scripts and a module to run a local web server so that we don't have to restart the whole server to see changes. We'll just need to refresh the page to get the new code on the client side. Eventually we'll have live reloading but I'm just trying to keep it simple for now.